Consolidated JsonPatchDocument extensions into JsonPatchDocumentOfT overloads, updated relevant tests
